Fixed Update System
The update system now works in Arceus. The only problem now is the release.dart script; it downloads the last successful build run for upload. The only problem is that the run build version can be out of sync of the release.dart script. Need to figure out a different way of storing the version of the app that the release.dart script can modify, but not the user.