Large Changes
SKit has been refactored to be smaller, simpler, and allows any object with the type of SRoot to be added to the root of the file. SHeader is the only expection, as it should be the only header in the file. Constellation names can now have special characters in them, and still be able to be saved. Fixed a bug when extracting. 'makeCurrent' does not await inside the 'Star' class, which lead to missing data inside files and the extraction process is not finished. Added some SKit type checking as well.